Refactor code to reinforce null checks and use BehaviorSubject
The update incorporates null checks for the items in the ZoneComponent to prevent runtime errors. Also, it refactors the AppComponent to use BehaviorSubject for zones, sites, places, and layers. Alongside this, certain corrections in ZoneDefinitionService have been made to ensure filtered items are accurately reflected in data. A few aesthetic changes in the application-component HTML file for enhanced user experience are also included.